feat: separates trade history into its own widget and uses it in the trade bottom widget
